Added disabling of hover when delete is disabled
When the delete button is disabled the hover effect is disabled. The
hover effect is enabled when the delete button is enabled. This is done
by adding and removing the enabled-button class which effectively
darkens the button's border on hover.

The repeating code in toggleAllCheckboxes() and toggleDeleteButton()
should be reduced. Ideally toggleAllCheckboxes() should call
toggleDeleteButton() to toggle the delete button and its hover effect.