Attempted to use existing list method
Attempted to make table sortable by creating a new list object after the
table has been created. While clicking on the name heading worked, the
table entries were removed from the DOM instead of being sorted.

Added the class "name" to the link element instead of the name cell
since it is the link which contains the actual text to sort.

To get the name heading to act as a button, it was necessary to add the
list.js attributes first, followed by the "titled-column" class.

The problem is that it is difficult to determine at which point in the
code the table is created since most functions operate asynchronously.
First attempted to call the createNewList() at the end of
traverseTree(). This didn't work because traverseTree() returns multiple
times for each removed bookmark, not at the end of the entire tree
traversal. Then tried calling createNewList() at the end of the callback
function of the chrome.bookmarks.getTree() method, but this didn't work
either because createNewList() is called before the video responses
arrive. Perhaps this is the correct point at which the tree traversal is
completed, it's just that the table isn't created yet because the table
entries are added in asynchronous functions. (i.e. the rows aren't created
until the responses arrive)