Added object-oriented pattern to requesting videos
Made sendRequest() a method of the object MatchedBookmark. The new
objects are created in scanNode(), then their sendRequest() methods are
called after the tree traversal in getBookmarks().

Attempted to define the methods sendRequest() and addToTable() by
binding them to the MatchedBookmark prototype, but there were issues
trying to call addToTable() from within sendRequest(). Hence, the
methods were defined from within the constructor function and the value
of "this" was saved to "self" which was then used to refer to the object
from within sendRequest().

Even though the object-oriented design is more elegant than the function
centered design, it still does not fix the problem of attempting to
create the List object before the table is completed.