Added sorting the table by the bookmark's name
Made the sorting functionality work by creating a new List object after
every table entry is added

Removed the createNewList() function and added its instructions to the
addToTable() method since it is was only called from addToTable() and
provided minimal functionality

While the sorting is working, it is only possible to sort the names of
the bookmarks in descending order. Once the name heading is clicked, its
"sort" class is changed to "sort desc". As a result, the user is not
able to switch between ascending and descending order.