Fixed sorting not reversing
Since the List object was being created after every addition to the
table, multiple event listeners were binded to the column heading which
were conflicting with each other.

The solution is to create the List object once the table has been
completed. This results in only one event listener for the column
heading. Now sorting the name column works as expected. That is, once
the column is sorted by name in ascending order, the next click on the
heading sorts the column in descending order, and so on.