Added Query object
Added the Query object which contains the group of video IDs to check,
the request for that group, and the response. This way when the response
returns a random amount of time later, it is paired with its
corresponding request, thus making it easy to determine which videos
have been removed.

Removed unrelated code of previous features which will be re-added later.